WikiFSSearch: split portable types, guard macOS-only impls
C1: Split portable Tantivy value types (TantivyDocumentKind,
    TantivyContentSnapshot, TantivyContentSource,
    TantivyShadowSearchResult) into new TantivySearchTypes.swift (no
    import TantivySwift — portable). TantivySearchDocument.swift keeps
    only the @TantivyDocument macro type behind #if os(macOS).
C2: Guard NLEmbedder.swift, TantivyIndexer.swift,
    TantivySearchService.swift with #if os(macOS).
C3: EmbeddingService — selectedEmbedderIdentifier() returns
    "unavailable-linux" on non-macOS; configure() is a no-op;
    isAvailable stays false. The store already handles
    isAvailable == false (cosine leg skipped).
C4: VectorCosine.dot — pure-Swift fallback when Accelerate is
    unavailable (identical result, no SIMD). Dormant on Linux since
    embeddings are unavailable anyway.

macOS build + full test suite (3257 tests) pass.
